Moved UploadScript out of ScriptService and into ScriptUploader - Pulled upload related tests from script_service_testrb to script_uploader_test.rb
