+++ lib/project_types/script/cli.rb
@@ -52,6 +52,7 @@ module Script
       autoload :ExtensionPointRepository, Project.project_filepath("layers/infrastructure/extension_point_repository")
       autoload :ScriptProjectRepository, Project.project_filepath("layers/infrastructure/script_project_repository")
       autoload :ScriptService, Project.project_filepath("layers/infrastructure/script_service")
+      autoload :ScriptUploader, Project.project_filepath("layers/infrastructure/script_uploader")
 
       module Languages
         autoload :AssemblyScriptProjectCreator,
