+++ lib/project_types/script/layers/application/push_script.rb
@@ -20,7 +20,9 @@ module Script
                 compiled_type: task_runner.compiled_type,
                 metadata: task_runner.metadata,
               )
-              uuid = Infrastructure::ScriptService.new(ctx: p_ctx, api_key: script_project.api_key).push(
+              script_service = Infrastructure::ScriptService.new(ctx: p_ctx, api_key: script_project.api_key)
+              module_upload_url = Infrastructure::ScriptUploader.new(script_service).upload(package.script_content)
+              uuid = script_service.push(
                 uuid: package.uuid,
                 extension_point_type: package.extension_point_type,
                 script_content: package.script_content,
@@ -28,6 +30,7 @@ module Script
                 force: force,
                 metadata: package.metadata,
                 script_json: package.script_json,
+                module_upload_url: module_upload_url,
               )
               script_project_repo.update_env(uuid: uuid)
               spinner.update_title(p_ctx.message("script.application.pushed"))
