+++ test/project_types/script/layers/infrastructure/script_service_test.rb
@@ -130,7 +130,6 @@ describe Script::Layers::Infrastructure::ScriptService do
         },
         resp: response
       )
-      Script::Layers::Infrastructure::ScriptService::UploadScript.any_instance.stubs(:call).returns(url)
     end
 
     subject do
@@ -142,9 +141,9 @@ describe Script::Layers::Infrastructure::ScriptService do
           schema_minor_version,
           use_msgpack,
         ),
-        script_content: script_content,
         script_json: script_json,
         api_key: api_key,
+        module_upload_url: url
       )
     end
 
@@ -281,105 +280,6 @@ describe Script::Layers::Infrastructure::ScriptService do
 
         it "should raise EmptyResponseError error" do
           assert_raises(Script::Layers::Infrastructure::Errors::EmptyResponseError) { subject }
-        end
-      end
-    end
-  end
-
-  describe "UploadScript" do
-    let(:instance) { Script::Layers::Infrastructure::ScriptService::UploadScript.new(script_service) }
-    subject { instance.call(api_key, script_content) }
-
-    let(:api_key) { "fake_key" }
-    let(:script_content) { "(module)" }
-    let(:module_upload_url_generate) do
-      <<~HERE
-        mutation moduleUploadUrlGenerate {
-          moduleUploadUrlGenerate {
-            url
-            userErrors {
-              field
-              message
-            }
-          }
-        }
-        HERE
-    end
-    let(:url) { "https://some-bucket" }
-    let(:response) do
-      {
-        data: {
-          scriptServiceProxy: JSON.dump(script_service_response),
-        },
-      }
-    end
-
-    before do
-      stub_load_query("script_service_proxy", script_service_proxy)
-      stub_load_query("module_upload_url_generate", module_upload_url_generate)
-      stub_partner_req(
-        "script_service_proxy",
-        variables: {
-          api_key: api_key,
-          variables: {}.to_json,
-          query: module_upload_url_generate,
-        },
-        resp: response
-      )
-    end
-
-    describe "when fail to apply module upload url" do
-      let(:script_service_response) do
-        {
-          "data" => {
-            "moduleUploadUrlGenerate" => {
-              "url" => nil,
-              "userErrors" => [{ "message" => "invalid", "field" => "appKey", "tag" => "user_error" }],
-            },
-          },
-        }
-      end
-
-      it "should raise GraphqlError" do
-        assert_raises(Script::Layers::Infrastructure::Errors::GraphqlError) { subject }
-      end
-    end
-
-    describe "when succeed to apply module upload url" do
-      let(:script_service_response) do
-        {
-          "data" => {
-            "moduleUploadUrlGenerate" => {
-              "url" => url,
-              "userErrors" => [],
-            },
-          },
-        }
-      end
-
-      describe "when fail to upload module" do
-        before do
-          stub_request(:put, url).with(
-            headers: { "Content-Type" => "application/wasm" },
-            body: script_content
-          ).to_return(status: 500)
-        end
-
-        it "should raise an ScriptUploadError" do
-          assert_raises(Script::Layers::Infrastructure::Errors::ScriptUploadError) { subject }
-        end
-      end
-
-      describe "when succeed to upload module" do
-        before do
-          stub_request(:put, url).with(
-            headers: { "Content-Type" => "application/wasm" },
-            body: script_content
-          ).to_return(status: 200)
-        end
-
-        it "should return the url" do
-          assert_equal(url, subject)
         end
       end
     end
